New_Version_Python/requirements.txt
customtkinter>=5.2.0
reportlab>=4.0.0
bcrypt>=4.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
python-barcode>=0.14.0
pandas>=2.0.0
openpyxl>=3.1.0
